Fix completion of envdir names when editing a complete command line
This fixes completion in the case below, when trying to complete "bar":

    envd foo bar<TAB> -- some-command …

by only looking for "--" in the part of the line up to the completion
point (the <TAB>).

Previously, the presence of "--" anywhere in the line caused normal
command/arg completion to always be done.  This bug got in the way of
re-editing command lines to add a forgotten dir or otherwise modify a
history entry.